requirements.txt
biopython==1.70
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
MySQL-python==1.2.5
numpy==1.14.0
pandas==0.22.0
PyMySQL==0.8.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
Werkzeug==0.14.1
